// mirinay-frontend/src/enums/EstadoEnum.ts
enum EstadoEnum {
    REMOVIDO= "REMOVIDO",
    DESCLASSIFICADO = "DESCLASSIFICADO",
    NAOENCONTRADO = "NAOENCONTRADO",
    DESAPROVADO = "DESAPROVADO",
    APROVADO = "APROVADO",
    COTANDO = "COTANDO",
    PREAPROVADO = "PREAPROVADO",
    A_PEDIR = "A_PEDIR",
    EMTRANSITO = "EMTRANSITO",
    ENTREGUE = "ENTREGUE",

}

export default EstadoEnum;